Remove unnecessary checks
In case of `found_last_byte == true` `bytes` is `&buf[0..=index]`,
where `index > 0` (because otherwise we cannot enter to match).
So `bytes.len()` is always > 0 and that means that `read` also > 0
and moreover it is the last byte of `bytes` here. Because `bytes`
is a view into `buf`, read is also a position of `>` in `buf`.